refactor(analysis): extract content nesting depth
